requirements.txt
langchain
llama_index
transformers
torch
python-dotenv
markdown2
docutils